More precise T_i error calc., code opt.
A method was created to measure the line intensity with a Gaussian fit. Another method was added that calculates the T_i error more precisely using the minimize function of scipy. Several modifications were made on the T_i fit function that made it three times faster.  The instrumental functios now are downloaded to the modul, because loading them from there is much faster than from the archive.